+++ benchmarks/generator/dataset/synthetic_prompt_generator.py
@@ -1,0 +1,164 @@
+import random
+
+# A collection of realistic text templates for generating prompts
+REALISTIC_TEMPLATES = [
+    # Question-answering templates
+    "Can you explain how {topic} works in simple terms?",
+    "What are the main differences between {topic_a} and {topic_b}?",
+    "I need to understand {topic} for my {purpose}. Can you help?",
+    "Could you provide a step-by-step guide on how to {action}?",
+    "What are the best practices for {activity} in {field}?",
+    
+    # Creative writing templates
+    "Write a short story about {character} who discovers {object} in {location}.",
+    "Create a poem about {theme} using the style of {author}.",
+    "Describe a scene where {character_a} meets {character_b} at {location}.",
+    "Write a dialogue between {character_a} and {character_b} discussing {topic}.",
+    "Develop a plot outline for a story about {theme} set in {setting}.",
+    
+    # Professional content templates
+    "Draft an email to {recipient} regarding {subject}.",
+    "Write a product description for {product} highlighting its {feature}.",
+    "Create a marketing copy for {service} targeting {audience}.",
+    "Compose a social media post announcing {event} for {platform}.",
+    "Draft a professional bio for {person} who specializes in {expertise}.",
+    
+    # Information retrieval templates
+    "Summarize the key points about {topic} in bullet points.",
+    "What are the latest developments in {field} as of 2024?",
+    "Provide a comparison table of {item_a}, {item_b}, and {item_c}.",
+    "What are the pros and cons of {subject}?",
+    "Give me 5 tips for improving {skill}."
+]
+
+# Domain-specific vocabulary to make the prompts more realistic
+TOPICS = [
+    "machine learning", "artificial intelligence", "neural networks", "deep learning", 
+    "natural language processing", "computer vision", "reinforcement learning",
+    "blockchain", "cryptocurrency", "smart contracts", "decentralized finance",
+    "cloud computing", "serverless architecture", "microservices", "containerization",
+    "cybersecurity", "ethical hacking", "network security", "encryption",
+    "data science", "big data", "data visualization", "statistical analysis",
+    "software development", "agile methodology", "DevOps", "continuous integration"
+]
+
+ACTIONS = [
+    "deploy a machine learning model", "optimize database queries", "secure a web application",
+    "build a responsive website", "create a mobile app", "implement an API",
+    "analyze data using Python", "set up a cloud infrastructure", "configure a firewall",
+    "develop a recommendation system", "train a neural network", "perform sentiment analysis"
+]
+
+CHARACTERS = [
+    "a software engineer", "a data scientist", "a startup founder", "a cybersecurity expert",
+    "an AI researcher", "a product manager", "a UX designer", "a digital nomad",
+    "a tech entrepreneur", "a blockchain developer", "a virtual reality designer"
+]
+
+LOCATIONS = [
+    "Silicon Valley", "a tech conference", "a coworking space", "a virtual reality world",
+    "a futuristic city", "a remote island with high-speed internet", "a hackathon",
+    "an innovation lab", "a digital marketplace", "an AI research center"
+]
+
+
+PADDING_PROMPT = [
+                f" Additionally, I'm interested in learning about {random.choice(TOPICS)}.",
+                f" Could you also explain how this relates to {random.choice(TOPICS)}?",
+                f" I'm asking because I need to {random.choice(ACTIONS)} for {random.choice(['my work', 'a client', 'a project', 'my research'])}.",
+                f" For context, I have experience with {random.choice(TOPICS)} but I'm new to this specific area.",
+                f" I've been trying to understand this concept for {random.choice(['days', 'weeks', 'months'])} and would appreciate a clear explanation."
+            ]
+
+def generate_synthetic_prompt(tokenizer, target_token_length):
+    """
+    Generate a realistic prompt using templates and domain-specific vocabulary
+    
+    Args:
+        tokenizer: The tokenizer to use
+        target_token_length: Desired length in tokens
+        
+    Returns:
+        A realistic prompt string
+    """
+    # Start with a random template
+    template = random.choice(REALISTIC_TEMPLATES)
+    
+    # Fill in the template with random relevant content
+    filled_template = template.format(
+        topic=random.choice(TOPICS),
+        topic_a=random.choice(TOPICS),
+        topic_b=random.choice(TOPICS),
+        purpose=random.choice(["project", "research", "presentation", "startup idea", "blog post"]),
+        action=random.choice(ACTIONS),
+        activity=random.choice(["coding", "designing", "analyzing", "implementing", "testing"]),
+        field=random.choice(["tech", "finance", "healthcare", "education", "e-commerce"]),
+        character=random.choice(CHARACTERS),
+        character_a=random.choice(CHARACTERS),
+        character_b=random.choice(CHARACTERS),
+        object=random.choice(["a quantum computer", "an AI assistant", "a time machine", "a virtual reality device"]),
+        location=random.choice(LOCATIONS),
+        theme=random.choice(["innovation", "digital transformation", "future of work", "technological singularity"]),
+        author=random.choice(["a tech visionary", "a sci-fi writer", "a futurist", "a digital artist"]),
+        setting=random.choice(["a smart city", "a space colony", "a digital universe", "a post-AI world"]),
+        recipient=random.choice(["a potential client", "a team member", "a project stakeholder", "a tech investor"]),
+        subject=random.choice(["project proposal", "software update", "partnership opportunity", "technical issue"]),
+        product=random.choice(["AI software", "smart device", "cloud service", "tech gadget"]),
+        feature=random.choice(["innovative features", "user-friendly interface", "cutting-edge technology", "performance"]),
+        service=random.choice(["consulting service", "tech solution", "software as a service", "digital platform"]),
+        audience=random.choice(["tech enthusiasts", "business professionals", "developers", "startups"]),
+        event=random.choice(["product launch", "tech conference", "software release", "hackathon"]),
+        platform=random.choice(["LinkedIn", "Twitter", "Facebook", "Instagram"]),
+        person=random.choice(CHARACTERS),
+        expertise=random.choice(TOPICS),
+        item_a=random.choice(TOPICS),
+        item_b=random.choice(TOPICS),
+        item_c=random.choice(TOPICS),
+        skill=random.choice(["programming", "data analysis", "system design", "technical writing", "debugging"])
+    )
+    
+    # Check token length
+    token_count = len(tokenizer.encode(filled_template))
+    
+    # If the template is too short, extend it with additional relevant content
+    while token_count < target_token_length:
+        # Add more content to the prompt
+        additional_content = [
+            f" Additionally, I'm interested in learning about {random.choice(TOPICS)}.",
+            f" Could you also explain how this relates to {random.choice(TOPICS)}?",
+            f" I'm asking because I need to {random.choice(ACTIONS)} for {random.choice(['my work', 'a client', 'a project', 'my research'])}.",
+            f" For context, I have experience with {random.choice(TOPICS)} but I'm new to this specific area.",
+            f" I've been trying to understand this concept for {random.choice(['days', 'weeks', 'months'])} and would appreciate a clear explanation."
+        ]
+        
+        filled_template += random.choice(additional_content)
+        token_count = len(tokenizer.encode(filled_template))
+    
+    # If the prompt is too long, truncate it to the desired length
+    if token_count > target_token_length:
+        tokenized = tokenizer.encode(filled_template)[:target_token_length]
+        filled_template = tokenizer.decode(tokenized, skip_special_tokens=True)
+    
+    return filled_template
+
+
+def adjust_prompt_length(tokenizer, prompt, target_token_length):
+    token_count = len(tokenizer.encode(prompt))
+    adjusted_prompt = prompt
+    if token_count < target_token_length:
+        while token_count < target_token_length:
+            additional_content = [
+                f" Additionally, I'm interested in learning about {random.choice(TOPICS)}.",
+                f" Could you also explain how this relates to {random.choice(TOPICS)}?",
+                f" I'm asking because I need to {random.choice(ACTIONS)} for {random.choice(['my work', 'a client', 'a project', 'my research'])}.",
+                f" For context, I have experience with {random.choice(TOPICS)} but I'm new to this specific area.",
+                f" I've been trying to understand this concept for {random.choice(['days', 'weeks', 'months'])} and would appreciate a clear explanation."
+            ]
+            adjusted_prompt += random.choice(additional_content)
+            token_count = len(tokenizer.encode(adjusted_prompt))
+    elif token_count > target_token_length:
+        adjusted_prompt_tokenized = tokenizer.encode(prompt)[:target_token_length]
+        adjusted_prompt = tokenizer.decode(adjusted_prompt_tokenized, skip_special_tokens=True)
+    return adjusted_prompt
+    
+
